fix: decimalfield para dinheiro

+++ apollo_lojinhaREST/Produtos/models.py
@@ -4,7 +4,7 @@ class Produtos(models.Model):
     '''Aqui será cadastrado o produto'''
     nomeProduto      = models.CharField(max_length=200, blank= False, verbose_name='Nome')
     descricaoProduto = models.TextField(verbose_name='Descrição')
-    precoProduto     = models.FloatField(blank=False, verbose_name='Preço')
+    precoProduto     = models.DecimalField(blank=False, verbose_name='Preço')
 
     def __str__(self):
         return f'{self.nomeProduto}'
